chore: 도메인 패키지 구조와 테스트 패키지 구조가 일치하도록 변경
