+++ src/test/java/domain/user/CardPoolTest.java
@@ -1,9 +1,8 @@
-package domain;
+package domain.user;
 
 import domain.card.Card;
 import domain.card.CardNumber;
 import domain.card.CardType;
-import domain.user.CardPool;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
