+++ src/test/java/domain/user/DealerTest.java
@@ -1,10 +1,8 @@
-package domain;
+package domain.user;
 
 import domain.card.Card;
 import domain.card.CardNumber;
 import domain.card.CardType;
-import domain.user.CardPool;
-import domain.user.Dealer;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
